@@ -5,11 +5,11 @@
   "proxy": "http://localhost:7071/",
   "dependencies": {
     "@apidevtools/json-schema-ref-parser": "^9.0.9",
-    "@craco/craco": "^6.4.0",
-    "@fluentui/react": "^8.41.0",
+    "@craco/craco": "^6.4.3",
+    "@fluentui/react": "^8.51.1",
     "@radiantearth/stac-fields": "1.0.0-beta.7",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@testing-library/jest-dom": "^5.15.0",
+    "@reduxjs/toolkit": "^1.7.1",
+    "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "@turf/bbox-polygon": "^6.5.0",
@@ -17,39 +17,39 @@
     "@turf/centroid": "^6.5.0",
     "@turf/union": "^6.5.0",
     "@types/geojson": "^7946.0.8",
-    "@types/jest": "^27.0.2",
+    "@types/jest": "^27.4.0",
     "@types/lodash-es": "^4.17.5",
-    "@types/marked": "^3.0.2",
+    "@types/marked": "^4.0.1",
     "@types/node": "^16.11.6",
-    "@types/react": "^17.0.34",
+    "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.20",
-    "@types/react-router-dom": "^5.3.2",
+    "@types/react-redux": "^7.1.22",
+    "@types/react-router-dom": "^5.3.3",
     "@uifabric/icons": "^7.6.2",
-    "axios": "^0.24.0",
-    "azure-maps-control": "^2.1.6",
+    "axios": "^0.25.0",
+    "azure-maps-control": "^2.1.8",
     "dayjs": "^1.10.7",
-    "dompurify": "^2.3.3",
+    "dompurify": "^2.3.5",
     "formik": "^2.2.9",
-    "highlight.js": "^11.3.1",
+    "highlight.js": "^11.4.0",
     "json-stringify-pretty-compact": "^3.0.0",
     "lodash-es": "^4.17.21",
-    "marked": "^3.0.4",
+    "marked": "^4.0.12",
     "number-abbreviate": "^2.0.0",
-    "query-string": "^7.0.1",
+    "query-string": "^7.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-error-boundary": "^3.1.4",
     "react-helmet": "^6.1.0",
-    "react-query": "^3.31.0",
+    "react-query": "^3.34.12",
     "react-redux": "^7.2.6",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.2.1",
     "react-router-hash-link": "^2.4.3",
     "react-scripts": "4.0.3",
-    "react-use": "^17.3.1",
+    "react-use": "^17.3.2",
     "redux": "^4.1.2",
-    "swagger-ui-react": "^4.0.1",
-    "typescript": "^4.4.4",
+    "swagger-ui-react": "^4.4.0",
+    "typescript": "^4.5.5",
     "yup": "^0.32.11"
   },
   "scripts": {
