@@ -1,12 +1,11 @@
 import React, { Suspense } from "react";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import { initializeIcons } from "@uifabric/icons";
 
 import Catalog from "./pages/Catalog";
 import Collection from "./pages/Collection";
 import NotFound from "./pages/NotFound";
 import AccountSurvey from "./pages/AccountSurvey";
-import Docs from "./pages/Docs";
 import Applications from "./pages/Applications";
 import { usePrefetchContent } from "./utils/requests";
 import Home from "./pages/Home";
@@ -17,6 +16,7 @@ import Layout from "./components/Layout";
 import { registerCustomIcons } from "utils/icons";
 
 const Explore = React.lazy(() => import("./pages/Explore"));
+const Docs = React.lazy(() => import("./pages/Docs"));
 
 function App() {
   initializeIcons(undefined, { disableWarnings: true });
@@ -31,43 +31,33 @@ function App() {
   return (
     <Router basename={process.env.PUBLIC_URL}>
       <div>
-        <Switch>
-          <Route exact path="/">
-            <Home />
-          </Route>
-          <Route exact path="/terms">
-            <Terms />
-          </Route>
-          <Route exact path="/catalog">
-            <Catalog />
-          </Route>
-          <Route path="/docs/">
-            <Docs />
-          </Route>
-          <Route path="/applications">
-            <Applications />
-          </Route>
-          <Route path="/dataset/group/:groupId">
-            <CatalogGroup />
-          </Route>
-          <Route path="/dataset/:id">
-            <Collection />
-          </Route>
-          <Route path="/account/request">
-            <AccountSurvey />
-          </Route>
-          <Route path="/explore">
-            <Suspense fallback={pageFallback}>
-              <Explore />
-            </Suspense>
-          </Route>
-          <Route path="/404">
-            <NotFound />
-          </Route>
-          <Route>
-            <NotFound />
-          </Route>
-        </Switch>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/terms" element={<Terms />} />
+          <Route path="/catalog" element={<Catalog />} />
+          <Route
+            path="/docs/*"
+            element={
+              <Suspense fallback={pageFallback}>
+                <Docs />
+              </Suspense>
+            }
+          />
+          <Route path="/applications" element={<Applications />} />
+          <Route path="/dataset/group/:groupId" element={<CatalogGroup />} />
+          <Route path="/dataset/:id" element={<Collection />} />
+          <Route path="/account/request" element={<AccountSurvey />} />
+          <Route
+            path="/explore"
+            element={
+              <Suspense fallback={pageFallback}>
+                <Explore />
+              </Suspense>
+            }
+          />
+          <Route path="/404" element={<NotFound />} />
+          <Route path="*" element={<NotFound />} />
+        </Routes>
       </div>
     </Router>
   );
